These ACL resources are duplicated.
I'm not sure how this ever worked. I was unable to log in as an admin
user, as Zend was throwing an exception about these resources already
existing. 'default:addemployeleaves' was merged in e4fb891afb for
example. I suspect that a bad git merge caused these, as git blame
references 2.1 and 3.2 commits.

Signed-Off-By: Rob Thomas <rthomas@sangoma.com>